Advanced-RAG/pyproject.toml
[tool.poetry]
name = "advanced-rag"
version = "0.1.0"
description = ""
authors = ["teruyo14 <teruyo.sui4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
langchainhub = "^0.1.21"
langchain-chroma = "^0.1.4"
langchain = "^0.3.3"
langgraph = "^0.2.35"
langchain-community = "^0.3.2"
langchain-openai = "^0.2.2"
google-search-results = "^2.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
